+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.30] - 2025-10-31
+
+### Changed
+- Refactored MarkdownRenderer to use marked library for improved markdown parsing
+- Renamed MarkdownRenderer to MarkdownService for clarity
+
+### Added
+- MarkedAdapter module to integrate marked library with application
+- Test suites for MarkdownService and MarkedAdapter
+
 ## [0.0.29] - 2025-10-30
 
 ### Added